Models, added auto increment

+++ main/models.py
@@ -19,7 +19,7 @@ class News(models.Model):
 
 
 class World(models.Model):
-    world_id = models.SmallIntegerField(primary_key=True)
+    world_id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=45, blank=True, null=True)
     name_value = models.CharField(max_length=45, blank=True, null=True)
     pvp_type = models.CharField(max_length=30, blank=True, null=True)
